refactor(f-flow): move plugin internals under namespaces
Move layout plugin sources under plugins/layout and snapping helpers under plugins/snapping.
Update internal imports, lint rules, and architecture docs for the new slice layout.
