--- libs/f-flow/src/f-layout/models/i-f-layout-algorithm-options.ts
@@ -1,4 +1,0 @@
-import { IFLayoutOptions } from './i-f-layout-options';
-
-export interface IFLayoutAlgorithmOptions<TAlgorithm extends string = string>
-  extends IFLayoutOptions<TAlgorithm> {}
