--- libs/f-flow/src/f-line-alignment/providers.ts
@@ -1,5 +1,0 @@
-import { FLineAlignmentComponent } from './f-line-alignment.component';
-
-export const F_LINE_ALIGNMENT_PROVIDERS = [
-    FLineAlignmentComponent,
-];
