+++ libs/f-flow/src/plugins/layout/f-layout/models/i-f-layout-algorithm-options.ts
@@ -1,0 +1,5 @@
+import { IFLayoutOptions } from './i-f-layout-options';
+
+export interface IFLayoutAlgorithmOptions<
+  TAlgorithm extends string = string,
+> extends IFLayoutOptions<TAlgorithm> {}
