+++ libs/f-flow/src/plugins/layout/f-layout/models/i-f-layout-node.ts
