+++ libs/f-flow/src/plugins/snapping/f-line-alignment/providers.ts
@@ -1,0 +1,3 @@
+import { FLineAlignmentComponent } from './f-line-alignment.component';
+
+export const F_LINE_ALIGNMENT_PROVIDERS = [FLineAlignmentComponent];
