+++ libs/f-flow/src/plugins/snapping/f-magnetic-rects/f-magnetic-rects-base.ts
