+++ libs/f-flow/src/plugins/snapping/f-magnetic-rects/providers.ts
